유저 페이지 조회 API 추가 및 구조 변경
