+++ src/main/java/NS/pgmg/dto/userpage/FindUserPageRequestDto.java
@@ -1,0 +1,17 @@
+package NS.pgmg.dto.userpage;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class FindUserPageRequestDto {
+
+    private String name;
+
+    @Builder
+    public FindUserPageRequestDto(String name) {
+        this.name = name;
+    }
+}
