+++ src/main/java/NS/pgmg/service/UserService.java
@@ -1,9 +1,14 @@
 package NS.pgmg.service;
 
+import NS.pgmg.config.JwtConfig;
 import NS.pgmg.domain.user.User;
+import NS.pgmg.dto.login.LoginDto;
+import NS.pgmg.dto.login.SocialRegisterAndLoginDto;
 import NS.pgmg.dto.register.BasicRegisterDto;
 import NS.pgmg.dto.register.ModelRegisterDto;
 import NS.pgmg.dto.register.ProPhotoRegisterDto;
+import NS.pgmg.dto.userpage.FindUserPageRequestDto;
+import NS.pgmg.dto.userpage.FindUserPageResponseDto;
 import NS.pgmg.exception.EmailDuplicateException;
 import NS.pgmg.exception.NameDuplicateException;
 import NS.pgmg.exception.PasswordMismatchException;
@@ -20,12 +25,11 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Map;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class UserRegisterService {
+public class UserService {
 
     private final UserRepository userRepository;
 
@@ -90,6 +94,54 @@ public class UserRegisterService {
         userRepository.save(findUser);
         log.info("기사 회원가입 완료, email = {}", request.getEmail());
         return new ResponseEntity<>("기사 회원가입이 완료되었습니다.", HttpStatus.CREATED);
+    }
+
+    @Transactional(readOnly = true)
+    public ResponseEntity<String> login(LoginDto request) {
+
+        String requestEmail = request.getEmail();
+        User findUser = userRepository.findByEmail(requestEmail);
+
+        if (!request.getEmail().equals(findUser.getEmail())) {
+            return new ResponseEntity<>("이메일이 올바르지 않습니다.", HttpStatus.BAD_REQUEST);
+        } else if (!request.getPasswd().equals(findUser.getPasswd())) {
+            return new ResponseEntity<>("비밀번호가 올바르지 않습니다.", HttpStatus.BAD_REQUEST);
+        } else {
+            log.info("로그인 이메일 = {}", requestEmail);
+            return new ResponseEntity<>(JwtConfig.getToken(requestEmail), HttpStatus.OK);
+        }
+    }
+
+    @Transactional
+    public ResponseEntity<String> socialLogin(SocialRegisterAndLoginDto request) {
+
+        String requestEmail = request.getEmail();
+        User findUser = userRepository.findByEmail(requestEmail);
+
+        if (findUser == null) {
+            User user = User.builder()
+                    .email(requestEmail)
+                    .socialTF(true)
+                    .build();
+            userRepository.save(user);
+            return new ResponseEntity<>(JwtConfig.getToken(requestEmail), HttpStatus.OK);
+        } else if (findUser.isSocialTF()) {
+            return new ResponseEntity<>(JwtConfig.getToken(requestEmail), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("일반 유저이메일입니다.", HttpStatus.BAD_REQUEST);
+        }
+    }
+
+    public FindUserPageResponseDto findUserPage(FindUserPageRequestDto request) {
+
+        User findUser = userRepository.findByName(request.getName());
+
+        return FindUserPageResponseDto.builder()
+                .email(findUser.getEmail())
+                .name(findUser.getName())
+                .baseUser("일반유저")
+                .intro(findUser.getIntro())
+                .build();
     }
 
     public void passwdValidation(String passwd, String passwd2) {
